Revert Crear producto tienda in main

+++ app/src/main/java/com/example/calitour/model/entity/User.kt
@@ -9,7 +9,6 @@ data class User (
     var email: String,
     var name : String,
     var phoneNumber: String,
-    var points : Long,
+    var points : Int,
     var photoUri: Uri
 )
-
